% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_datamap_to_influxdb.R
\name{datamap_to_body}
\alias{datamap_to_body}
\title{This function creates the string for the raw-body
   used by the influxdb to transfer the datamap}
\usage{
datamap_to_body(rv, type = "raw")
}
\arguments{
\item{rv}{The global rv object.}

\item{type}{(String) The type of the output. E.g. "raw" or "json"}
}
\description{
This function creates the string for the raw-body
   used by the influxdb to transfer the datamap
}
