% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{feedback}
\alias{feedback}
\title{Simple method to feedback messages either to the user and/or to the console.}
\usage{
feedback(
  print_this,
  type = "Info",
  ui = FALSE,
  console = TRUE,
  prefix = "",
  suffix = "",
  findme = ""
)
}
\arguments{
\item{print_this:}{The string to be showed}

\item{type:}{(Optional) The type of message as string.
This is also the headline of the modal.
If type is e.g. "Warning"
the printed line will be "[Warning] print_this".}

\item{ui:}{(Optional) If true, the message will also be printed
to the user in form of a modal.}

\item{console:}{(Optional) If true, the message will also be printed
to the console as is.}

\item{prefix:}{Prefix (String)}

\item{suffix:}{Suffix (String)}

\item{findme:}{(Optional) String to find the message in the code.
E.g. 10-digit random hex from https://www.browserling.com/tools/random-hex}
}
\description{
Extended version of the printme-function.
}
