% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_db_settings}
\alias{get_db_settings}
\title{get_db_settings}
\usage{
get_db_settings(input, target = T)
}
\arguments{
\item{input}{Shiny server input object}

\item{target}{A boolean (default: TRUE).}
}
\description{
get_db_settings
}
