% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchApp}
\alias{launchApp}
\title{Launch the DQA graphical user interface (GUI)}
\usage{
launchApp(port = 3838, utilspath, db_source)
}
\arguments{
\item{port}{The port, DQAgui is running on (default: 3838)}
}
\value{
DQAgui Shiny application
}
\description{
Launch the DQA graphical user interface (GUI)
}
