% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCompleteness.R
\name{module_completeness_ui}
\alias{module_completeness_ui}
\title{module_completeness_ui}
\usage{
module_completeness_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_completeness_ui
}
