% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleConfig2.R
\name{module_config2_ui}
\alias{module_config2_ui}
\title{module_config_ui}
\usage{
module_config2_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_config_ui
}
