% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDashboard.R
\name{module_dashboard_ui}
\alias{module_dashboard_ui}
\title{module_dashboard_ui}
\usage{
module_dashboard_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_dashboard_ui
}
