% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleReport.R
\name{module_report_ui}
\alias{module_report_ui}
\title{module_report_ui}
\usage{
module_report_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_report_ui
}
