% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleUniquePlausibility.R
\name{module_uniq_plaus_ui}
\alias{module_uniq_plaus_ui}
\title{module_uniq_plaus_ui}
\usage{
module_uniq_plaus_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_uniq_plaus_ui
}
