% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleVisualizations.R
\name{module_visualizations_ui}
\alias{module_visualizations_ui}
\title{module_visualizations_ui}
\usage{
module_visualizations_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_visualizations_ui
}
