% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCompleteness.R
\name{moduleCompletenessUI}
\alias{moduleCompletenessUI}
\title{moduleCompletenessUI}
\usage{
moduleCompletenessUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleCompletenessUI
}
