% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{on_start}
\alias{on_start}
\title{on_start}
\usage{
on_start(session, rv, input, output)
}
\arguments{
\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{input}{Shiny server input object}

\item{output}{Shiny server output object}
}
\description{
on_start
}
