% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{printme}
\alias{printme}
\title{Simply prints stuff to the console.}
\usage{
printme(print_this, type = "Info")
}
\arguments{
\item{print_this}{The string to be printed.}

\item{type}{(Optional) The type of message as string.
If type is e.g. "Warning" the printed line will be "[Warning] print_this".}
}
\description{
Simply prints stuff to the console.
}
