% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{This function checks if all necessary input parameters
  for source and target exist and are valid.}
\usage{
validate_inputs(rv)
}
\arguments{
\item{rv}{The global 'reactiveValues()' object, defined in server.R}
}
\description{
This function checks if all necessary input parameters
  for source and target exist and are valid.
}
